Solve Lens . At composition with a separate Lens#at method
Enabling composition for lenses and At instances requires access to the output type of
the lens, which is not available in the ComposeLens interface. Lens#at is convenient for most cases and where you really need to compose a lens with an At you can use

```typescript
lens.compose(at("foo").toLens())
```

which will infer the output type contextually.